+++ extensions/bing-wallpaper/src/utils/constants.ts
@@ -1,0 +1,6 @@
+export enum WallpaperTag {
+  ALL = "All",
+  ONLINE = "Online",
+  DOWNLOADED = "Downloaded",
+}
+export const wallpaperTags = Object.entries(WallpaperTag);