+++ extensions/bing-wallpaper/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Bing Wallpaper Changelog
 
+## [Add downloaded wallpaper list] - 2022-05-20
+
+- Add downloaded wallpaper list in Set Bing Wallpaper command
+- Random Bing Wallpaper will randomize to the downloaded wallpapers
+
 ## [Fix text errors] - 2022-04-26
 
 - Fix text errors